Extracted benchmark code into common module.
